rebase/move new_user_queryable into core
