+++ python/src/opendp/prelude.py
@@ -7,5 +7,5 @@ from opendp.metrics import *
 from opendp.measures import *
 from opendp.typing import *
 from opendp.accuracy import *
-from opendp.core import new_function
+from opendp.core import new_function, new_user_queryable
 from opendp.context import *
